Refactored the previous reveals and renamed them so they fall into either one of the two categories: `Render` or `Color`. I've also created a new script called `ColorTypeWriter.cs` that essentially does what `RenderTypeWriter.cs` did but instead of stopping characters from rendering it will always render each characters mesh but reveal it by using color instead. I've also added comments to all the scripts so you can understand how each reveal is done.
Render Reveal: Characters aren't even being displayed by TMP until we tell it to. (This means that the mesh for each character isn't being generated/calculated)

Color Reveal: Characters are being rendered by TMP but are hidden by changing the alpha of each character mesh to zero.

RenderTypeWriter.cs: The way we are revealing each character is by incrementing the maxVisibleCharacters int on the TextMeshProUGUI class.

ColorTypeWriter.cs & ColorRandomWriter.cs: The way we are revealing each character is by altering the color of the vertices on the characters mesh. Each character mesh has 4 vertices and we are changing the color32 of each corner at once.